0.19.10 new package for Skills + refactoring
