+++ wfrp/src/main/java/com/warhammer/wfrp/skills/SkillDto.java
@@ -1,4 +1,4 @@
-package com.warhammer.wfrp.dto.skill;
+package com.warhammer.wfrp.skills;
 
 public record SkillDto(String name,
                        String attribute,
